Use a Breadth-first search (BFS) algorithm in order to crawl the websites
Instead of using a recursive function and an array to keep track of the websites which need to be crawled, now it uses a queue in an iterative fashion. There seemed to be a bug in previous version where an item was removed from the array before your links were extracted.